sm_database inherits from ResidualsCompliantRecordSet
