+++ smtk/rcrs.py
@@ -40,6 +40,9 @@ class ResidualsCompliantRecordSet:
     used in :meth:`gmpe_residuals.Residuals.get_residuals` to compute the
     records residuals. Subclasses need to implement few abstract-like methods
     defining how to get the input data required for the residuals computation
+    For subclassing examples, see:
+    :class:`smtk.sm_database.GroundMotionDatabase`
+    :class:`smtk.sm_table.GroundMotionTable`
     '''
 
     # SCALAR_IMTS = ["PGA", "PGV", "PGD", "CAV", "Ia"]
@@ -300,217 +303,6 @@ class ResidualsCompliantRecordSet:
         '''
         for imtx, values in observations.items():
             observations[imtx] = np.asarray(values, dtype=float)
-
-
-class GroundMotionDatabase(ResidualsCompliantRecordSet):
-
-    def __init__(self, db_id, db_name, db_directory=None, records=[],
-                 site_ids=[]):
-        """
-        """
-        self.id = db_id
-        self.name = db_name
-        self.directory = db_directory
-        self._records = [rec for rec in records]
-        self.site_ids = site_ids
-
-    ###########################################################
-    # IMPLEMENTS ResidualsCompliantRecordSet ABSTRACT METHODS #
-    ###########################################################
-
-    @property
-    def records(self):
-        '''Returns an iterable of records from this database (e.g. list, tuple,
-        generator). Note that as any iterable, the returned object does not
-        need to define a length whereby `len(self.records)` will work: this
-        will depend on subclasses implementation
-        '''
-        return self._records
-
-    def get_record_eventid(self, record):
-        '''Returns the record event id (usually int) of the given record'''
-        return record.event.id
-
-    def update_sites_context(self, record, context):
-        '''Updates the attributes of `context` with the given `record` data.
-        `context` is a :class:`openquake.hazardlib.contexts.SitesContext`
-        object. In the typical implementation it has the attributes defined in
-        `self.sites_context_attrs` all initialized to empty lists.
-        Here you should append to those attributes the relative record value,
-        e.g.: `context.vs30.append(record.vs30)`
-        '''
-        context.vs30.append(record.site.vs30)
-        context.lons.append(record.site.longitude)
-        context.lats.append(record.site.latitude)
-        if record.site.altitude:
-            depth = record.site.altitude * -1.0E-3
-        else:
-            depth = 0.0
-        context.depths.append(depth)
-        if record.site.vs30_measured is not None:
-            vs30_measured = record.site.vs30_measured
-        else:
-            vs30_measured = 0
-        context.vs30measured.append(vs30_measured)
-        if record.site.z1pt0 is not None:
-            z1pt0 = record.site.z1pt0
-        else:
-            z1pt0 = vs30_to_z1pt0_cy14(record.site.vs30)
-        context.z1pt0.append(z1pt0)
-        if record.site.z2pt5 is not None:
-            z2pt5 = record.site.z2pt5
-        else:
-            z2pt5 = vs30_to_z2pt5_cb14(record.site.vs30)
-        context.z2pt5.append(z2pt5)
-        if getattr(record.site, "backarc", None) is not None:
-            context.backarc.append(record.site.backarc)
-
-    def update_distances_context(self, record, context):
-        '''Updates the attributes of `context` with the given `record` data.
-        `context` is a :class:`openquake.hazardlib.contexts.DistancesContext`
-        object. In the typical implementation it has the attributes defined in
-        `self.distances_context_attrs` all initialized to empty lists.
-        Here you should append to those attributes the relative record value,
-        e.g.: `context.rjb.append(record.rjb)`
-        '''
-        context.repi.append(record.distance.repi)
-        context.rhypo.append(record.distance.rhypo)
-        # TODO Setting Rjb == Repi and Rrup == Rhypo when missing value
-        # is a hack! Need feedback on how to fix
-        if record.distance.rjb is not None:
-            context.rjb.append(record.distance.rjb)
-        else:
-            context.rjb.append(record.distance.repi)
-        if record.distance.rrup is not None:
-            context.rrup.append(record.distance.rrup)
-        else:
-            context.rrup.append(record.distance.rhypo)
-        if record.distance.r_x is not None:
-            context.rx.append(record.distance.r_x)
-        else:
-            context.rx.append(record.distance.repi)
-        if getattr(record.distance, "ry0", None) is not None:
-            context.ry0.append(record.distance.ry0)
-        if getattr(record.distance, "rcdpp", None) is not None:
-            context.rcdpp.append(record.distance.rcdpp)
-        if record.distance.azimuth is not None:
-            context.azimuth.append(record.distance.azimuth)
-        if record.distance.hanging_wall is not None:
-            context.hanging_wall.append(record.distance.hanging_wall)
-        if getattr(record.distance, "rvolc", None) is not None:
-            context.rvolc.append(record.distance.rvolc)
-
-    def update_rupture_context(self, record, context, nodal_plane_index=1):
-        '''Updates the attributes of `context` with the given `record` data.
-        `context` is a :class:`openquake.hazardlib.contexts.RuptureContext`
-        object. In the typical implementation it has the attributes defined in
-        `self.sites_context_attrs` all initialized to NaN (numpy.nan).
-        Here you should set those attributes with the relative record value,
-        e.g.: `context.mag = record.event.mag`
-        '''
-        context.mag = record.event.magnitude.value
-        if nodal_plane_index == 2:
-            context.strike = \
-                record.event.mechanism.nodal_planes.nodal_plane_2['strike']
-            context.dip = \
-                record.event.mechanism.nodal_planes.nodal_plane_2['dip']
-            context.rake = \
-                record.event.mechanism.nodal_planes.nodal_plane_2['rake']
-        elif nodal_plane_index == 1:
-            context.strike = \
-                record.event.mechanism.nodal_planes.nodal_plane_1['strike']
-            context.dip = \
-                record.event.mechanism.nodal_planes.nodal_plane_1['dip']
-            context.rake = \
-                record.event.mechanism.nodal_planes.nodal_plane_1['rake']
-        else:
-            context.strike = 0.0
-            context.dip = 90.0
-            context.rake = \
-                record.event.mechanism.get_rake_from_mechanism_type()
-
-        if record.event.rupture.surface:
-            context.ztor = record.event.rupture.surface.get_top_edge_depth()
-            context.width = record.event.rupture.surface.width
-            context.hypo_loc = \
-                record.event.rupture.surface.get_hypo_location(1000)
-        else:
-            if record.event.rupture.depth is not None:
-                context.ztor = record.event.rupture.depth
-            else:
-                context.ztor = record.event.depth
-
-            if record.event.rupture.width is not None:
-                context.width = record.event.rupture.width
-            else:
-                # Use the PeerMSR to define the area and assuming an aspect ratio
-                # of 1 get the width
-                context.width = \
-                    np.sqrt(sm_utils.DEFAULT_MSR.get_median_area(context.mag,
-                                                                 0))
-
-            # Default hypocentre location to the middle of the rupture
-            context.hypo_loc = (0.5, 0.5)
-        context.hypo_depth = record.event.depth
-        context.hypo_lat = record.event.latitude
-        context.hypo_lon = record.event.longitude
-
-    def update_observations(self, record, observations, component="Geometric"):
-        '''Updates the observed intensity measures types (imt) with the given
-        `record` data. `observations` is a `dict` of imts (string) mapped to
-        numeric lists. Here you should append to each list the imt value
-        derived from `record`, ususally numeric or NaN (`numpy.nan`):
-        ```
-            for imt, values in observations.items():
-                if imtx in self.SCALAR_IMTS:  # currently, 'PGA' or 'PGV'
-                    val = ... get the imt scalar value from record ...
-                elif "SA(" in imtx:
-                    val = ... get the SA numeric array / list from record ...
-                else:
-                    raise ValueError("IMT %s is unsupported!" % imtx)
-                values.append(val)
-        ```
-        '''
-        selection_string = "IMS/H/Spectra/Response/Acceleration/"
-        fle = h5py.File(record.datafile, "r")
-        for imtx in self.imts:
-            if imtx in self.SCALAR_IMTS:
-                observations[imtx].append(
-                    self.get_scalar(fle, imtx, component))
-            elif "SA(" in imtx:
-                target_period = imt.from_string(imtx).period
-                spectrum = fle[selection_string + component +
-                               "/damping_05"].value
-                periods = fle["IMS/H/Spectra/Response/Periods"].value
-                observations[imtx].append(get_interpolated_period(
-                    target_period, periods, spectrum))
-            else:
-                raise ValueError("IMT %s is unsupported!" % imtx)
-        fle.close()
-
-    ###########################
-    # END OF ABSTRACT METHODS #
-    ###########################
-
-    def get_scalar(self, fle, i_m, component="Geometric"):
-        """
-        Retrieves the scalar IM from the database
-        :param fle:
-            Instance of :class: h5py.File
-        :param str i_m:
-            Intensity measure
-        :param str component:
-            Horizontal component of IM
-        """
-        if not ("H" in fle["IMS"].keys()):
-            x_im = fle["IMS/X/Scalar/" + i_m].value[0]
-            y_im = fle["IMS/Y/Scalar/" + i_m].value[0]
-            return SCALAR_XY[component](x_im, y_im)
-        else:
-            if i_m in fle["IMS/H/Scalar"].keys():
-                return fle["IMS/H/Scalar/" + i_m].value[0]
-            else:
-                raise ValueError("Scalar IM %s not in record database" % i_m)
 
 
 class GroundMotionTable(ResidualsCompliantRecordSet):
